// src/Enums/RouteEnums.ts
export enum Route {

    //Client 
    CLIENT_ME =  "me",
    CLIENT_BASE =  "/clients/",
    TICKET_CODE_CLIENT = "/clients/tickets/all",
    TICKET_CODE_WINNER = "/ticket-codes/all_win",

    // Giveaway
    GIVEAWAY_BASE = "/giveaways/",

    // Adress
    ADDRESS_BASE = "/adresss",

    // Game

    GAME_BASE = "/games",

    // Employe
    EMPLOYE_ME = "me",
    EMPLOYE_BASE = "/employees/",


    // Admin
    ADMIN_ME = "me",
    ADMIN_BASE  = "/admins/",


    // Ticketcode
    TICKET_CODE_BASE = "/ticket-codes/",
    TICKET_CODE_APPLY = "/ticket-codes/apply/",
    


    //Image 
    IMAGE_BASE =  "/image",

}